Reorganize root-level types into feature folders
