+++ src/EventLogExpert.UI/DependencyInjection/UiServiceCollectionExtensions.cs
@@ -13,7 +13,7 @@ public static class UiServiceCollectionExtensions
         /// <summary>Registers the EventLogExpert UI-library services into the host container.</summary>
         /// <remarks>
         ///     The registered <see cref="KeyboardShortcutService" /> resolves <c>IMenuActionService</c>,
-        ///     <c>IModalCoordinator</c>, and <c>ISettingsService</c> from the container — the host must register those
+        ///     <c>IModalCoordinator</c>, and <c>ISettingsService</c> from the container - the host must register those
         ///     abstractions (the Runtime layer via <c>AddEventLogRuntime</c> plus the host's menu adapter) before resolving UI
         ///     services.
         /// </remarks>
